add check to removechsh.sh

+++ files/scripts/removechsh.sh
@@ -3,4 +3,8 @@
 # Tell build process to exit if there are any errors.
 set -oue pipefail
 
-rm /usr/bin/chsh
+if [ ! -f /usr/bin/chsh ]; then
+  echo "chsh not found in /usr/bin"
+else
+  rm /usr/bin/chsh
+fi
